BINIUS-640: Merge the extension module into the field module
The extension module held one trait and its degree-1 blanket impl, both
written entirely in terms of Field. Nothing in it needed a file of its own.

Its 116 lines move into field.rs, after Field and FieldOps. The crate-root
re-export of field widens to carry the trait, so binius_field::ExtensionField
keeps working and no crate outside binius-field changes.
