+++ crates/field/src/fields/ghash_sq.rs
@@ -29,7 +29,7 @@ use crate::{
 	Field, Ghash128b,
 	arch::{M128, M256, m256_from_u128s},
 	binary_field::{BinaryField, BinaryField1b, binary_field, impl_field_extension},
-	extension::ExtensionField,
+	field::ExtensionField,
 	underlier::U1,
 };
