+++ crates/field/src/lib.rs
@@ -16,7 +16,6 @@ pub mod arch;
 pub mod arithmetic_traits;
 pub mod binary_field;
 mod divisible;
-pub mod extension;
 pub mod field;
 pub mod fields;
 pub mod linear_transformation;
@@ -38,8 +37,7 @@ pub mod util;
 pub use arithmetic_traits::WideMul;
 pub use binary_field::*;
 pub use divisible::Divisible;
-pub use extension::*;
-pub use field::{Field, FieldOps};
+pub use field::{ExtensionField, Field, FieldOps};
 pub use fields::{ghash::*, ghash_sq::*, rijndael::*};
 pub use maskable::Maskable;
 pub use packed::PackedField;
